Fixes cyclic import warning
