+++ src/app/(pages)/calendar.tsx
@@ -4,7 +4,7 @@ import Divider from '@/components/Elements/Universal/Divider'
 import ErrorView from '@/components/Elements/Universal/ErrorView'
 import ToggleRow from '@/components/Elements/Universal/ToggleRow'
 import { type Colors } from '@/components/colors'
-import { UserKindContext } from '@/components/provider'
+import { UserKindContext } from '@/components/contexts'
 import { useRefreshByUser } from '@/hooks'
 import { USER_GUEST } from '@/hooks/contexts/userKind'
 import { guestError, networkError } from '@/utils/api-utils'
