+++ src/app/(tabs)/food.tsx
@@ -2,7 +2,7 @@ import { MealDay } from '@/components/Elements/Food'
 import ErrorView from '@/components/Elements/Universal/ErrorView'
 import PlatformIcon from '@/components/Elements/Universal/Icon'
 import { type Colors } from '@/components/colors'
-import { FoodFilterContext } from '@/components/provider'
+import { FoodFilterContext } from '@/components/contexts'
 import { useRefreshByUser } from '@/hooks'
 import { type Food } from '@/types/neuland-api'
 import { networkError } from '@/utils/api-utils'
