+++ src/app/(timetable)/details.tsx
@@ -8,7 +8,7 @@ import FormList from '@/components/Elements/Universal/FormList'
 import PlatformIcon, { chevronIcon } from '@/components/Elements/Universal/Icon'
 import ShareButton from '@/components/Elements/Universal/ShareButton'
 import { type Colors } from '@/components/colors'
-import { NotificationContext, RouteParamsContext } from '@/components/provider'
+import { NotificationContext, RouteParamsContext } from '@/components/contexts'
 import { useNotification } from '@/hooks'
 import i18n, { type LanguageKey } from '@/localization/i18n'
 import { type FormListSections, type SectionGroup } from '@/types/components'
