+++ src/app/(user)/about.tsx
@@ -3,7 +3,7 @@ import { chevronIcon, linkIcon } from '@/components/Elements/Universal/Icon'
 import SectionView from '@/components/Elements/Universal/SectionsView'
 import SingleSectionPicker from '@/components/Elements/Universal/SingleSectionPicker'
 import { type Colors } from '@/components/colors'
-import { AppIconContext, FlowContext } from '@/components/provider'
+import { AppIconContext, FlowContext } from '@/components/contexts'
 import { type FormListSections } from '@/types/components'
 import { PAGE_BOTTOM_SAFE_AREA, PAGE_PADDING } from '@/utils/style-utils'
 import { useTheme } from '@react-navigation/native'
