+++ src/app/(user)/appicon.tsx
@@ -2,7 +2,7 @@ import Divider from '@/components/Elements/Universal/Divider'
 import PlatformIcon from '@/components/Elements/Universal/Icon'
 import SectionView from '@/components/Elements/Universal/SectionsView'
 import { type Colors } from '@/components/colors'
-import { AppIconContext } from '@/components/provider'
+import { AppIconContext } from '@/components/contexts'
 import { capitalizeFirstLetter } from '@/utils/app-utils'
 import { getStatusBarStyle } from '@/utils/ui-utils'
 import { useTheme } from '@react-navigation/native'
