+++ src/app/(user)/dashboard.tsx
@@ -2,7 +2,7 @@ import Divider from '@/components/Elements/Universal/Divider'
 import PlatformIcon from '@/components/Elements/Universal/Icon'
 import { type Card, type ExtendedCard } from '@/components/allCards'
 import { type Colors } from '@/components/colors'
-import { DashboardContext, UserKindContext } from '@/components/provider'
+import { DashboardContext, UserKindContext } from '@/components/contexts'
 import { getDefaultDashboardOrder } from '@/hooks/contexts/dashboard'
 import { USER_GUEST } from '@/hooks/contexts/userKind'
 import { arraysEqual } from '@/utils/app-utils'
