+++ src/components/Cards/LibraryCard.tsx
@@ -2,7 +2,7 @@ import API from '@/api/authenticated-api'
 import { NoSessionError } from '@/api/thi-session-handler'
 import Divider from '@/components/Elements/Universal/Divider'
 import { type Colors } from '@/components/colors'
-import { FlowContext, UserKindContext } from '@/components/provider'
+import { FlowContext, UserKindContext } from '@/components/contexts'
 import { type Reservation } from '@/types/thi-api'
 import { formatFriendlyDateTimeRange } from '@/utils/date-utils'
 import { LoadingState } from '@/utils/ui-utils'
