+++ src/components/Cards/TimetableCard.tsx
@@ -1,6 +1,6 @@
 import Divider from '@/components/Elements/Universal/Divider'
 import { type Colors } from '@/components/colors'
-import { UserKindContext } from '@/components/provider'
+import { UserKindContext } from '@/components/contexts'
 import { useInterval } from '@/hooks/useInterval'
 import { type FriendlyTimetableEntry } from '@/types/utils'
 import { formatFriendlyDateTime, formatFriendlyTime } from '@/utils/date-utils'
