+++ src/components/provider.tsx
@@ -1,12 +1,5 @@
 import { useAppState, useOnlineManager } from '@/hooks'
-import { type AppIconHook } from '@/hooks/contexts/appIcon'
-import { type FlowHook } from '@/hooks/contexts/flow'
-import { type IdCard } from '@/hooks/contexts/idCard'
-import {
-    DEFAULT_TIMETABLE_MODE,
-    type TimetableHook,
-    useTimetable,
-} from '@/hooks/contexts/timetable'
+import { useTimetable } from '@/hooks/contexts/timetable'
 import i18n from '@/localization/i18n'
 import { trackEvent } from '@aptabase/react-native'
 import AsyncStorage from '@react-native-async-storage/async-storage'
@@ -19,7 +12,7 @@ import { createAsyncStoragePersister } from '@tanstack/query-async-storage-persi
 import { QueryClient, focusManager } from '@tanstack/react-query'
 import { PersistQueryClientProvider } from '@tanstack/react-query-persist-client'
 import { usePathname } from 'expo-router'
-import React, { createContext, useEffect } from 'react'
+import React, { useEffect } from 'react'
 import {
     type AppStateStatus,
     Platform,
@@ -39,14 +32,20 @@ import {
     useTheme,
     useUserKind,
 } from '../hooks/contexts'
-import { type Dashboard } from '../hooks/contexts/dashboard'
-import { type FoodFilter } from '../hooks/contexts/foodFilter'
-import {
-    type Notifications,
-    useNotifications,
-} from '../hooks/contexts/notifications'
-import { type RouteParams } from '../hooks/contexts/routing'
+import { useNotifications } from '../hooks/contexts/notifications'
 import { type AppTheme, accentColors, darkColors, lightColors } from './colors'
+import {
+    AppIconContext,
+    DashboardContext,
+    FlowContext,
+    FoodFilterContext,
+    IdCardContext,
+    NotificationContext,
+    RouteParamsContext,
+    ThemeContext,
+    TimetableContext,
+    UserKindContext,
+} from './contexts'
 
 interface ProviderProps {
     children: React.ReactNode
@@ -69,92 +68,6 @@ export const queryClient = new QueryClient({
 
 const asyncStoragePersister = createAsyncStoragePersister({
     storage: AsyncStorage,
-})
-
-export const RouteParamsContext = createContext<RouteParams>({
-    routeParams: '',
-    updateRouteParams: () => {},
-    lecture: null,
-    updateLecture: () => {},
-})
-
-export const FoodFilterContext = createContext<FoodFilter>({
-    allergenSelection: [],
-    preferencesSelection: [],
-    selectedRestaurants: [],
-    showStatic: false,
-    foodLanguage: 'default',
-    toggleSelectedAllergens: () => {},
-    initAllergenSelection: () => {},
-    toggleSelectedPreferences: () => {},
-    toggleSelectedRestaurant: () => {},
-    toggleShowStatic: () => {},
-    toggleFoodLanguage: () => {},
-})
-
-export const UserKindContext = createContext<any>({
-    userKind: 'student',
-    userFaculty: 'unknown',
-    userFullName: '',
-    toggleUserKind: () => {},
-    updateUserFullName: () => {},
-})
-
-export const ThemeContext = createContext<any>({
-    accentColor: 'blue',
-    toggleAccentColor: () => {},
-})
-
-export const AppIconContext = createContext<AppIconHook>({
-    appIcon: 'default',
-    unlockedAppIcons: [],
-    toggleAppIcon: () => {},
-    addUnlockedAppIcon: () => {},
-})
-
-export const DashboardContext = createContext<Dashboard>({
-    shownDashboardEntries: [],
-    hiddenDashboardEntries: [],
-    hideDashboardEntry: () => {},
-    bringBackDashboardEntry: () => {},
-    resetOrder: () => {},
-    updateDashboardOrder: () => {},
-})
-
-export const FlowContext = createContext<FlowHook>({
-    isOnboarded: true,
-    toggleOnboarded: () => {},
-    isUpdated: true,
-    toggleUpdated: () => {},
-    analyticsAllowed: false,
-    toggleAnalytics: () => {},
-    analyticsInitialized: false,
-    initializeAnalytics: () => {},
-})
-
-export const ReloadProvider = createContext<any>({
-    reload: false,
-    toggleReload: () => {},
-})
-
-export const TimetableContext = createContext<TimetableHook>({
-    timetableMode: DEFAULT_TIMETABLE_MODE,
-    setTimetableMode: () => {},
-    selectedDate: new Date(),
-    setSelectedDate: () => {},
-})
-
-export const NotificationContext = createContext<Notifications>({
-    timetableNotifications: {},
-    updateTimetableNotifications: () => {},
-    deleteTimetableNotifications: () => {},
-    removeNotification: () => {},
-})
-
-export const IdCardContext = createContext<IdCard>({
-    mensaBalance: 0,
-    idLastUpdated: null,
-    updateMensaBalance: () => {},
 })
 
 /**
